+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -83,10 +83,10 @@ import Cardano.Ledger.Conway.Governance (
   epochStateDRepPulsingStateL,
   finishDRepPulser,
   gasDRepVotesL,
+  proposalsLookupId,
   psDRepDistrG,
   rsEnactStateL,
   setCompleteDRepPulsingState,
-  snapshotLookupId,
   utxosGovStateL,
   votingDRepThreshold,
   votingStakePoolThreshold,
@@ -459,7 +459,7 @@ lookupGovActionState ::
   ImpTestM era (Maybe (GovActionState era))
 lookupGovActionState aId = do
   proposals <- getsNES $ nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . proposalsGovStateL
-  pure $ snapshotLookupId aId proposals
+  pure $ proposalsLookupId aId proposals
 
 -- | Looks up the governance action state corresponding to the governance action id
 getGovActionState ::